Move remnant handling into remnants.go.
Keep Config, Extract, and callback wiring where they live; the action normalizer, snapshot/classify/clear helpers, and tests sit together.
